Extract addPRsToDoc
